app/src/main/java/com/example/remaster/NotificationHelper.java
package com.example.remaster;

import android.annotation.TargetApi;
import android.app.Notification;
import android.app.NotificationChannel;
import android.app.NotificationManager;
import android.content.Context;
import android.content.ContextWrapper;
import android.os.Build;

import androidx.core.app.NotificationCompat;


public class NotificationHelper extends ContextWrapper{
    public static final String channelID = "channelID";
    public static final String channelName = "Channel Name";
    public String contact;
    public String message;

    private NotificationManager nManager;

    public NotificationHelper(Context base){
        super(base);
        System.out.println("in Notification Helper");
        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
            createChannel();
        }
    }

    @TargetApi(Build.VERSION_CODES.O)
    public void createChannel() {
        NotificationChannel channel = new NotificationChannel(channelID, channelName, NotificationManager.IMPORTANCE_HIGH);
        channel.enableLights(true);
        channel.enableVibration(true);
        channel.setLightColor(R.color.colorPrimary);
        channel.setLockscreenVisibility(Notification.VISIBILITY_PRIVATE);
        getManager().createNotificationChannel(channel);
    }

    public NotificationManager getManager() {
        if (nManager == null){
            nManager = (NotificationManager) getSystemService(Context.NOTIFICATION_SERVICE);
        }

        return nManager;
    }

    public void setContact(String contact) {
        this.contact = contact;
    }

    public void setMessage(String message) {
        this.message = message;
    }

    public NotificationCompat.Builder getChannelNotification() {
        return new NotificationCompat.Builder(getApplicationContext(), channelID)
                .setContentTitle(contact)
                .setContentText(message)
                .setSmallIcon(R.drawable.ic_message);
    }
}
